🎨 moved some fiztures to global
